+++ eotsmanager/cmd/eotsd/daemon/pop_test.go
@@ -1,6 +1,12 @@
 package daemon_test
 
 import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/babylonlabs-io/finality-provider/eotsmanager/cmd/eotsd/daemon"
@@ -59,8 +65,34 @@ func TestPoPValidBabySignEotsPk(t *testing.T) {
 func TestPoPVerify(t *testing.T) {
 	t.Parallel()
 	for _, pop := range popsToVerify {
-		valid, err := daemon.VerifyPopExport(pop)
+		valid, err := daemon.ValidPopExport(pop)
 		require.NoError(t, err)
 		require.True(t, valid)
+	}
+}
+
+func TestPoPValidate(t *testing.T) {
+	t.Parallel()
+	validateCmd := daemon.NewPopValidateExportCmd()
+
+	tmp := t.TempDir()
+
+	for i, pop := range popsToVerify {
+		jsonString, err := json.MarshalIndent(pop, "", "  ")
+		require.NoError(t, err)
+
+		writer := bytes.NewBuffer([]byte{})
+		validateCmd.SetOutput(writer)
+
+		fileName := filepath.Join(tmp, fmt.Sprintf("%d-pop-out.json", i))
+		err = os.WriteFile(fileName, jsonString, 0644)
+		require.NoError(t, err)
+
+		validateCmd.SetArgs([]string{fileName})
+
+		err = validateCmd.ExecuteContext(context.Background())
+		require.NoError(t, err)
+
+		require.Equal(t, writer.String(), "Proof of Possession is valid!\n")
 	}
 }